fix(tests): fix problem when multiple tests are using one dbconnection
