+++ test/app.e2e-spec.ts
@@ -2,6 +2,8 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { AppModule } from './../src/app.module';
+import { PassTests } from './passTests';
+import { UserTest } from './userTests';
 
 describe('AppController (e2e)', () => {
   let app: INestApplication;
@@ -16,9 +18,9 @@ describe('AppController (e2e)', () => {
   });
 
   it('/ (GET)', () => {
-    return request(app.getHttpServer())
-      .get('/')
-      .expect(200)
-      .expect('Hello World!');
+    return request(app.getHttpServer()).get('/').expect(200).expect('Hello World!');
   });
+
+  UserTest();
+  PassTests();
 });
